feat(api): add user registration with organization setup
Replaces signin with register functionality, including creating
default organization, user-organization mapping, and default
project initialization during user registration.
